WIP: Split code into modules
closes #77
