+++ src/error.rs
@@ -1,0 +1,33 @@
+use std::{fmt, result};
+
+pub type Result<T> = result::Result<T, Error>;
+
+/// Custom error type
+pub enum Error {
+    /// Sourcecode could be located
+    SourceCodeNotFound,
+    /// Git is not installed or did not function properly
+    GitNotInstalled,
+    /// Did not find any git data in the directory
+    NoGitData,
+    /// An IO error occoured while reading ./
+    ReadDirectory,
+    /// Not in a Git Repo
+    NotGitRepo,
+    /// Error while getting branch info
+    ReferenceInfoError,
+}
+
+impl fmt::Debug for Error {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let content = match self {
+            Error::SourceCodeNotFound => "Could not find any source code in this directory",
+            Error::GitNotInstalled => "Git failed to execute",
+            Error::NoGitData => "Could not retrieve git configuration data",
+            Error::ReadDirectory => "Could not read directory",
+            Error::NotGitRepo => "This is not a Git Repo",
+            Error::ReferenceInfoError => "Error while retrieving reference information",
+        };
+        write!(f, "{}", content)
+    }
+}
